refactor: extract media library tooling
